+++ packages/apps/demo/src/Demo.tsx
@@ -1,14 +1,11 @@
-import { UnreachableError } from '@truckermudgeon/base/precon';
-import type { AtsSelectableDlc } from '@truckermudgeon/map/constants';
-import { AtsDlc, AtsSelectableDlcs } from '@truckermudgeon/map/constants';
+import { AtsSelectableDlcs } from '@truckermudgeon/map/constants';
 import {
+  allIcons,
   BaseMapStyle,
+  defaultMapStyle,
   GameMapStyle,
   MapIcon,
   SceneryTownSource,
-  StateCode,
-  allIcons,
-  defaultMapStyle,
 } from '@truckermudgeon/ui';
 import 'maplibre-gl/dist/maplibre-gl.css';
 import { useState } from 'react';
@@ -17,29 +14,9 @@ import MapGl, {
   FullscreenControl,
   NavigationControl,
 } from 'react-map-gl/maplibre';
-import { Legend } from './Legend';
+import { createListProps, Legend } from './Legend';
 import { MapSelectAndSearch } from './MapSelectAndSearch';
-
-const createListProps = <T,>(
-  selectedItems: Set<T>,
-  setSelectedItems: (value: React.SetStateAction<Set<T>>) => void,
-  allItems: ReadonlySet<T>,
-) => ({
-  selectedItems,
-  onSelectAllToggle: (all: boolean) =>
-    setSelectedItems(new Set(all ? allItems : [])),
-  onItemToggle: (item: T, newValue: boolean) => {
-    setSelectedItems((prevState: Set<T>) => {
-      const newState = new Set(prevState);
-      if (newValue) {
-        newState.add(item);
-      } else {
-        newState.delete(item);
-      }
-      return newState;
-    });
-  },
-});
+import { toStateCodes } from './state-codes';
 
 const Demo = () => {
   const [autoHide, setAutoHide] = useState(true);
@@ -116,42 +93,5 @@ const Demo = () => {
     </MapGl>
   );
 };
-
-function toStateCodes(atsDlcs: Set<AtsSelectableDlc>) {
-  return new Set<StateCode>([...atsDlcs].map(toStateCode).concat(StateCode.CA));
-}
-
-function toStateCode(atsDlc: AtsSelectableDlc): StateCode {
-  switch (atsDlc) {
-    case AtsDlc.Arizona:
-      return StateCode.AZ;
-    case AtsDlc.Colorado:
-      return StateCode.CO;
-    case AtsDlc.Idaho:
-      return StateCode.ID;
-    case AtsDlc.Kansas:
-      return StateCode.KS;
-    case AtsDlc.Montana:
-      return StateCode.MT;
-    case AtsDlc.Nevada:
-      return StateCode.NV;
-    case AtsDlc.NewMexico:
-      return StateCode.NM;
-    case AtsDlc.Oklahoma:
-      return StateCode.OK;
-    case AtsDlc.Oregon:
-      return StateCode.OR;
-    case AtsDlc.Texas:
-      return StateCode.TX;
-    case AtsDlc.Utah:
-      return StateCode.UT;
-    case AtsDlc.Washington:
-      return StateCode.WA;
-    case AtsDlc.Wyoming:
-      return StateCode.WY;
-    default:
-      throw new UnreachableError(atsDlc);
-  }
-}
 
 export default Demo;
